requirements.txt
absl-py==0.11.0
appdirs==1.4.4
applicationinsights==0.11.9
asgiref==3.3.1
astor==0.8.1
astroid==2.5.1
attrs==20.3.0
autopep8==1.5.5
backcall==0.2.0
backports.csv==1.0.7
bleach==3.3.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cli-helpers==0.2.3
Click==7.0
configobj==5.0.6
crcmod==1.7
decorator==4.4.2
distlib==0.3.1
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
enum34==1.1.10
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
Flask-PyMongo==2.3.0
future==0.18.2
gast==0.4.0
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
google-cloud-bigquery==2.10.0
google-cloud-core==1.6.0
google-cloud-datastore==2.1.0
google-cloud-language==2.0.0
google-cloud-logging==2.2.0
google-cloud-spanner==3.1.0
google-cloud-storage==1.36.1
google-cloud-translate==3.0.2
google-cloud-videointelligence==2.0.0
google-cloud-vision==2.2.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.36.1
h5py==3.2.0
httplib2==0.19.0
humanize==3.2.0
idna==2.10
importlib-metadata==3.7.0
ipython==7.21.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.5.2
libcst==0.3.17
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
mssql-cli==1.0.0
mypy==0.812
mypy-extensions==0.4.3
numpy==1.20.1
oauth2client==4.1.3
packaging==20.9
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pluggy==0.13.1
prompt-toolkit==3.0.16
proto-plus==1.14.2
protobuf==3.15.4
psycopg2==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycurl==7.43.0.2
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.8.0
PyGObject==3.30.4
PyJWT==2.0.1
pylint==2.7.2
pyls-mypy==0.1.8
pymongo==3.11.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-apt==1.8.4.3
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rope==0.18.0
rsa==4.7.2
six==1.15.0
snowballstemmer==2.1.0
sqlparse==0.4.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminaltables==3.1.0
tink==1.5.0
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.4.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.30.0
zipp==3.4.0
